feat: moved service folder to clicktopay folder
