+++ packages/lib/src/components/Card/components/ClickToPay/ClickToPayComponent.tsx
@@ -1,11 +1,11 @@
 import { Fragment, h } from 'preact';
 import { useEffect } from 'preact/hooks';
-import { CtpState } from '../../services/ClickToPayService';
+import { CtpState } from './services/ClickToPayService';
 import useClickToPayContext from './context/useClickToPayContext';
 import CtPOneTimePassword from './components/CtPOneTimePassword';
 import CtPCardsList from './components/CtPCardsList/CtPCardsList';
 import CtPSection from './components/CtPSection';
-import { CheckoutPayload } from '../../services/types';
+import { CheckoutPayload } from './services/types';
 
 type ClickToPayComponentProps = {
     onSubmit(payload: CheckoutPayload): void;
