+++ packages/lib/src/components/Card/components/ClickToPay/services/sdks/SrciError.ts
