+++ packages/lib/src/components/Card/types.ts
@@ -11,8 +11,8 @@ import {
     CbObjOnBinLookup
 } from '../internal/SecuredFields/lib/types';
 import { CVCPolicyType, DatePolicyType } from '../internal/SecuredFields/lib/types';
-import { IdentityLookupParams } from './services/types';
-import { SrcInitParams } from './services/sdks/types';
+import { IdentityLookupParams } from './components/ClickToPay/services/types';
+import { SrcInitParams } from './components/ClickToPay/services/sdks/types';
 
 export interface CardElementProps extends UIElementProps {
     /**
